auth_service/pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = ""
authors = ["Tani <vtvarma2012@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
fastapi = ">=0.121.2,<0.122.0"
uvicorn = ">=0.38.0,<0.39.0"
python-jose = {version=">=3.5.0,<4.0.0", extras=["cryptography"]}
passlib = {extras = ["argon2"], version = "^1.7.4"}
pydantic = ">=2.12.4,<3.0.0"
email-validator = ">=2.3.0,<3.0.0"
sqlalchemy = ">=2.0.44,<3.0.0"
psycopg2-binary = ">=2.9.11,<3.0.0"
python-multipart = ">=0.0.20,<0.0.21"
shared = {path = "../shared"}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
